fix missing places

+++ aten/src/ATen/TensorIterator.cpp
@@ -1406,9 +1406,9 @@ bool DimCounter::is_done() const {
 
 void DimCounter::increment(const std::array<int64_t, 2>& step) {
   offset += step[0] * step[1];
-  int ndim = values.size();
+  int64_t ndim = values.size();
   int64_t overflow = step[0];
-  int64_t i = 0;
+  int i = 0;
   if (step[1] != 1) {
     TORCH_INTERNAL_ASSERT(step[0] == shape[0] && values[0] == 0);
     i = 1;
